feat: don't require ua-parser-js in lambda-context
BREAKING CHANGE: If you want to continue to use `ua-parser-js` in
`lambda-context` you will need to use `lambdaContextWithUaParser` and
pass in your own parser function as described here:
https://github.com/birchill/bugsnag-zero?tab=readme-ov-file#using-a-custom-user-agent-string-parser

This is because from version `ua-parser-js` version 2 onwards, users are
required to either use the AGPL version or license the library.

It's best to allow users of this package to decide how they want to
consume `ua-parser-js`.
