+++ package.json
@@ -70,8 +70,5 @@
     "typescript": "5.6.3",
     "vitest": "2.1.5",
     "vitest-github-actions-reporter": "0.11.1"
-  },
-  "dependencies": {
-    "ua-parser-js": "^1.0.32"
   }
 }
